soldier death tweaks
DebriefingState::prepareDebriefing()
	- add mission XP even to dead or missing soldiers
GeoscapeState::time1Day()
	- rework chance to die today
GraphsState: touch for camelCase
BattleUnit::postMissionProcedures()
	- not a bool return
	- only calc recovery time if not deceased or MIA
		- cap recTime @100% instead of 150%
	- remove funct, updateGeoscapeStats()
Soldier
	- remove addMissionCount() & addKillCount()
	- replace w/ postTactical()
	- rename getWoundPCT() by getRecoveryPCT()

+++ src/Basescape/SoldierDiaryMissionState.cpp
@@ -118,7 +118,7 @@ SoldierDiaryMissionState::SoldierDiaryMissionState(
 		diary = deadSoldier->getDiary();
 
 		missionId = diary->getMissionIdList().at(_rowEntry);
-		if (missionId > static_cast<int>(missionStatistics->size()))
+		if (missionId > missionStatistics->size())
 			missionId = 0;
 
 		daysWounded = missionStatistics->at(missionId)->injuryList[deadSoldier->getId()];
@@ -140,7 +140,7 @@ SoldierDiaryMissionState::SoldierDiaryMissionState(
 		diary = soldier->getDiary();
 
 		missionId = diary->getMissionIdList().at(_rowEntry);
-		if (missionId > static_cast<int>(missionStatistics->size()))
+		if (missionId > missionStatistics->size())
 			missionId = 0;
 
 		daysWounded = missionStatistics->at(missionId)->injuryList[soldier->getId()];
